Fix return type of vector LengthSquared
LengthSquared now returns an integer for integer-based
vectors rather than double. This makes LengthSquared
consistent with DistanceSquared.

+++ src/Aardvark.Base.FSharp/Math/Vectors.fs
@@ -213,7 +213,7 @@ module Vec =
         let lengthSquaredWorking () =
             let a : float32 = lengthSquared V3f.One
             let a : float = lengthSquared V3d.One
-            let a : float = lengthSquared V3i.One
+            let a : int = lengthSquared V3i.One
             ()
 
         let normalizeWorking () =
